"Save Code" should only work with ILSpyTreeNodes.
